+++ src/main/java/ru/hse/learning_algorithm/TPMTrainer.java
@@ -11,18 +11,18 @@ public class TPMTrainer {
 
     private int epochs = 200;
 
-    public ArrayList<Integer>[] synchronize(TreeParityMachine tpm1, TreeParityMachine tpm2) {
+    public ArrayList<Short>[] synchronize(TreeParityMachine tpm1, TreeParityMachine tpm2) {
         if (tpm1.getLearningParadigm() != tpm2.getLearningParadigm())
             throw new NeuralNetException("Алгоритмы обучения не совпадают!");
-        int k = 0;
-        ArrayList<Integer> result = new ArrayList<>();
-        ArrayList<Integer> outputTPM1 = new ArrayList<>();
-        ArrayList<Integer> outputTPM2 = new ArrayList<>();
+        short k = 0;
+        ArrayList<Short> result = new ArrayList<>();
+        ArrayList<Short> outputTPM1 = new ArrayList<>();
+        ArrayList<Short> outputTPM2 = new ArrayList<>();
         int[] params = tpm1.getTPMParams();
-        int[] input = Random.getInts(params[0], -1, 1);
+        short[] input = Random.getInts(params[0], -1, 1);
         while (k < epochs) {
-            int out1 = tpm1.getOutput(input);
-            int out2 = tpm2.getOutput(input);
+            short out1 = tpm1.getOutput(input);
+            short out2 = tpm2.getOutput(input);
             if (Arrays.equals(tpm1.getSecretKey(), tpm2.getSecretKey())) {
                 result.add(k);
                 break;
@@ -42,8 +42,8 @@ public class TPMTrainer {
         return new ArrayList[]{outputTPM1, outputTPM2, result};
     }
 
-    public int synchronize(TreeParityMachine tpm1, int[] input, int out2) {
-        int out1 = tpm1.getOutput(input);
+    public short synchronize(TreeParityMachine tpm1, short[] input, short out2) {
+        short out1 = tpm1.getOutput(input);
         if (out1 == out2)
             tpm1.train(input, out2);
         return tpm1.getOutput(input);
