+++ src/main/java/ru/hse/learning_algorithm/Training.java
@@ -2,6 +2,6 @@ package ru.hse.learning_algorithm;
 
 public interface Training {
 
-    void train(int[] input, int output);
+    void train(short[] input, short output);
 
 }
